requirements.txt
streamlit
requests
PyPDF2
speechrecognition
fpdf
pillow
